requirements.txt
asgiref==3.7.2
Django==5.0
django-markdownx==4.0.7
Markdown==3.5.1
Pillow==10.1.0
sqlparse==0.4.4
uWSGI==2.0.23
